requirements.txt
Django>=3.1.12
djangorestframework==3.12.4
flake8==3.9.0
wikipedia-api
wikipedia~=1.4.0
pysbd~=0.3.4